requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (SQLite for development)
sqlalchemy==2.0.23
alembic==1.12.1

# Configuration and validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# File processing
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-multipart==0.0.6

# Machine Learning and NLP (optional)
transformers
torch
spacy
nltk
sentence-transformers
scikit-learn
sentencepiece

# Indonesian language processing
spacy-udpipe==1.0.0
regex==2023.10.3

# Text analysis and quality metrics
textstat==0.7.3
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Grammar and language checking
language-tool-python==2.7.1

# Progress tracking
tqdm==4.66.1

# PostgreSQL dependencies (install separately if needed)
# psycopg2-binary==2.9.9
# asyncpg==0.29.0
